计算repl log file size大小时没有乘以单位
git-svn-id: http://192.168.20.11/sequoiadb/trunk@12837 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/include/dpsReplicaLogMgr.hpp
@@ -168,7 +168,7 @@ namespace engine
       INT32 move( const DPS_LSN_OFFSET &offset,
                   const DPS_LSN_VER &version ) ;
 
-      void setLogFileSz ( UINT32 logFileSz )
+      void setLogFileSz ( UINT64 logFileSz )
       {
          _logger.setLogFileSz ( logFileSz ) ;
       }
